pyproject.toml
[project]
name = "docflow"
version = "0.1.0"
description = "Pipeline local de conversão, tradução e exportação de PDFs via Docling + Ollama"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "python-multipart>=0.0.20",
    "beautifulsoup4>=4.13",
    "python-docx>=1.1",
    "weasyprint>=63",
    "aiofiles>=24",
]

[project.scripts]
docflow = "backend.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "pytest-playwright>=0.7",
    "respx>=0.22",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
